+++ src/lib.rs
@@ -4,12 +4,21 @@
 //!
 //! ## Compatibility
 //!
-//! | crate | imgui | ash          | gpu-allocator (feature) | vk-mem (feature)                        |
-//! |-------|-------|--------------|-------------------------|-----------------------------------------|
-//! | 1.3.0 | 0.8   | [0.34, 0.37] | 0.18                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
-//! | 1.2.0 | 0.8   | 0.36         | 0.17                    | 0.2.3 ([forked][forked-mem-rs-036])     |
-//! | 1.1.x | 0.8   | 0.35         | 0.15                    | 0.2.3 ([forked][forked-mem-rs-035])     |
-//! | 1.0.0 | 0.8   | 0.35         | 0.14                    | 0.2.3 ([forked][forked-mem-rs-035])     |
+//! | crate  | imgui | ash          | gpu-allocator (feature) | vk-mem (feature)                        |
+//! |--------|-------|--------------|-------------------------|-----------------------------------------|
+//! | 1.12.0 | 0.11  | [0.34, 0.37] | 0.25                    | 0.3.0                                   |
+//! | 1.11.0 | 0.11  | [0.34, 0.37] | 0.25                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.10.0 | 0.11  | [0.34, 0.37] | 0.23                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.9.0  | 0.11  | [0.34, 0.37] | 0.22                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.8.0  | 0.10  | [0.34, 0.37] | 0.22                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.7.0  | 0.10  | [0.34, 0.37] | [0.19, 0.21]            | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.6.x  | 0.9   | [0.34, 0.37] | [0.19, 0.21]            | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.5.0  | 0.8   | [0.34, 0.37] | 0.19                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.4.0  | 0.8   | [0.34, 0.37] | 0.18                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.3.0  | 0.8   | [0.34, 0.37] | 0.18                    | 0.2.3 ([forked][forked-mem-rs-034-037]) |
+//! | 1.2.0  | 0.8   | 0.36         | 0.17                    | 0.2.3 ([forked][forked-mem-rs-036])     |
+//! | 1.1.x  | 0.8   | 0.35         | 0.15                    | 0.2.3 ([forked][forked-mem-rs-035])     |
+//! | 1.0.0  | 0.8   | 0.35         | 0.14                    | 0.2.3 ([forked][forked-mem-rs-035])     |
 //!
 //! ## How it works
 //!
@@ -49,6 +58,12 @@
 //!
 //! This feature adds support for [vk-mem-rs][vk-mem-rs]. It adds `Renderer::with_vk_mem_allocator` which takes
 //! a `Arc<Mutex<vk_mem::Allocator>>`. All internal allocator are then done using the allocator.
+//!
+//! ### dynamic-rendering
+//!
+//! This feature is useful if you want to integrate the library in an app making use of Vulkan's dynamic rendering.
+//! When enabled, functions that usually takes a `vk::RenderPass` as argument will now take a `vk::Format` which is
+//! the format of the color attachment the UI will be drawn to.
 //!
 //! ## Integration
 //!
@@ -94,16 +109,19 @@
 //! ## Example can be one of the following value:
 //! ## - collapsing_header
 //! ## - color_button
+//! ## - creating_windows
 //! ## - custom_textures
 //! ## - disablement
 //! ## - draw_list
 //! ## - hello_world
+//! ## - id_wrangling
 //! ## - keyboard
 //! ## - long_list
-//! ## - multiple fonts
+//! ## - long_table
+//! ## - multiple_fonts
 //! ## - progress_bar
 //! ## - radio_button
-//! ## - slider : TODO: crashes
+//! ## - slider
 //! ## - tables_api
 //! ## - test_drawing_channels_split
 //! ## - test_window_impl